requirements.txt
flask
python-dotenv
google-generativeai
pdfplumber
python-pptx
